+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 ### 10.2.0 Features
 
 - `[Cards]` Added Angular Wrapper Component for the IDS Cards. ([#1036](https://github.com/infor-design/enterprise-ng/issues/1036)) `EA`
+- `[Monthview]` Added Angular Wrapper Component for the IDS MonthView. ([#1051](https://github.com/infor-design/enterprise-ng/issues/1051)) `EA`
 
 ## v10.1.0